+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Added css handler class for image link wrapper
+
 ## [3.142.0] - 2021-03-09
 ### Added
 - Push `newsletterInput` pixel event when newsletter is mounted